Move initialize-store-service.js out of the instance-initializers directory.
That directory seems to be suggesting to ember-load-initializers that
it is a full initializer object and causing an exception to be thrown
in the globals build.

Fixes #5118
