+++ app/instance-initializers/ember-data.js
@@ -1,4 +1,4 @@
-import initializeStoreService from 'ember-data/instance-initializers/initialize-store-service';
+import initializeStoreService from 'ember-data/initialize-store-service';
 
 export default {
   name: "ember-data",
